--- a/Cabal/Distribution/FieldGrammar/Pretty.hs
+++ b/Cabal/Distribution/FieldGrammar/Pretty.hs
@@ -66,5 +66,5 @@ instance FieldGrammar PrettyFieldGrammar where
     knownField _           = pure ()
     deprecatedSince [] _ _ = PrettyFG (\_ -> mempty)
     deprecatedSince _  _ x = x
-    availableSince _       = id
+    availableSince _ _     = id
     hiddenField _          = PrettyFG (\_ -> mempty)