--- a/Cabal/tests/ParserTests/errors/leading-comma.cabal
+++ b/Cabal/tests/ParserTests/errors/leading-comma.cabal
@@ -1,0 +1,20 @@
+name:                leading-comma
+version:             0
+synopsis:            leading comma, trailing comma, or ordinary
+build-type:          Simple
+-- too small cabal-version
+cabal-version:       2.0
+
+library
+  default-language: Haskell2010
+  exposed-modules:  LeadingComma
+
+  build-depends: base, containers
+
+  build-depends:
+    deepseq,
+    transformers,
+
+  build-depends:
+    , filepath
+    , directory