--- a/Cabal/tests/ParserTests/errors/leading-comma.errors
+++ b/Cabal/tests/ParserTests/errors/leading-comma.errors
@@ -1,0 +1,1 @@
+PError (Position 16 18) "\nunexpected end of input\nexpecting white space: \"deepseq,\\ntransformers,\""