--- a/Cabal/tests/ParserTests/regressions/leading-comma.cabal
+++ b/Cabal/tests/ParserTests/regressions/leading-comma.cabal
@@ -1,0 +1,19 @@
+name:                leading-comma
+version:             0
+synopsis:            leading comma, trailing comma, or ordinary
+build-type:          Simple
+cabal-version:       2.1
+
+library
+  default-language: Haskell2010
+  exposed-modules:  LeadingComma
+
+  build-depends: base, containers
+
+  build-depends:
+    deepseq,
+    transformers,
+
+  build-depends:
+    , filepath
+    , directory