We need versioned grammars:

- `availableSince`, that would "fix" https://github.com/haskell/cabal/issues/4448 because backpack features won't be parsed. (User will get parse warnings about not-yet-supported spec version).
- SPDX license, how to parse `license:` field will change: https://github.com/haskell/cabal/issues/2547

Leading comma is just a simple extension to experiment with :)

I'm not 100% sure that `singletons`-ish approach is *the best*, but it has good features:
- we use dictionary as an argument, so the approach is at least as good (performance wise) as passing an argument:
  `Versioned v => ... -> a` vs. `Version -> ... -> a`.
- The dictionaries can be specialised (have to check the core if it actually works)
- Types are harder to change accidentally, a little like having `MonadReader Version => ... -> m a`.

It works, it's easy to see by changing `cabal-version` to `2.0` in `leading-comma.cabal`. Tests will fail. ~I'll add negative tests after common-stanza branch is merged.~ negative test added.

cc @23Skidoo @hvr

Surprisingly? this seems to be as performant as current `master`, parsing all Hackage packages starting with `a`: (including acme-everything :)

### this PR

```
 % /usr/bin/time -v .../parser-hackage-tests parse-parsec a +RTS -s
Reading index from: /home/ogre/.cabal/packages/hackage.haskell.org/01-index.tar
6714 files processed
  27,946,371,856 bytes allocated in the heap
   2,082,050,064 bytes copied during GC
       6,101,392 bytes maximum residency (970 sample(s))
       1,084,576 bytes maximum slop
              14 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     25627 colls,     0 par    2.304s   2.300s     0.0001s    0.0032s
  Gen  1       970 colls,     0 par    0.067s   0.067s     0.0001s    0.0009s

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time    4.228s  (  4.225s elapsed)
  GC      time    2.371s  (  2.366s elapsed)
  EXIT    time   -0.000s  ( -0.000s elapsed)
  Total   time    6.598s  (  6.591s elapsed)

  %GC     time      35.9%  (35.9% elapsed)

  Alloc rate    6,610,560,408 bytes per MUT second

  Productivity  64.1% of total user, 64.1% of total elapsed
```

### this PR without SPECIALIZED in D.PD.FieldGrammar

```
6714 files processed
  27,921,069,992 bytes allocated in the heap
   2,105,778,208 bytes copied during GC
       7,756,992 bytes maximum residency (983 sample(s))
         731,304 bytes maximum slop
              17 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     25473 colls,     0 par    2.275s   2.271s     0.0001s    0.0057s
  Gen  1       983 colls,     0 par    0.067s   0.067s     0.0001s    0.0009s

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time    4.188s  (  4.185s elapsed)
  GC      time    2.341s  (  2.338s elapsed)
  EXIT    time   -0.000s  ( -0.000s elapsed)
  Total   time    6.529s  (  6.522s elapsed)

  %GC     time      35.9%  (35.8% elapsed)

  Alloc rate    6,667,467,459 bytes per MUT second

  Productivity  64.1% of total user, 64.2% of total elapsed
```

### master

```
% /usr/bin/time -v .../parser-hackage-tests parse-parsec a +RTS -s
Reading index from: /home/ogre/.cabal/packages/hackage.haskell.org/01-index.tar
6714 files processed
  27,943,265,024 bytes allocated in the heap
   2,079,495,768 bytes copied during GC
       7,922,800 bytes maximum residency (974 sample(s))
         396,456 bytes maximum slop
              17 MB total memory in use (0 MB lost due to fragmentation)

                                     Tot time (elapsed)  Avg pause  Max pause
  Gen  0     25618 colls,     0 par    2.246s   2.242s     0.0001s    0.0049s
  Gen  1       974 colls,     0 par    0.065s   0.065s     0.0001s    0.0010s

  INIT    time    0.000s  (  0.000s elapsed)
  MUT     time    4.277s  (  4.272s elapsed)
  GC      time    2.311s  (  2.307s elapsed)
  EXIT    time   -0.001s  ( -0.001s elapsed)
  Total   time    6.587s  (  6.578s elapsed)

  %GC     time      35.1%  (35.1% elapsed)

  Alloc rate    6,533,837,864 bytes per MUT second

  Productivity  64.9% of total user, 64.9% of total elapsed
```

---

leading comma for `CommaFSep`/`CommaVSep` fields, i.e. fields with mandatory comma are (atm):

- build-depends
- build-tool-depends
- build-tools
- mixins
- pkgconfig-depends
- reexported-modules
- setup-depends

---

Please include the following checklist in your PR:

* [x] Patches conform to the [coding conventions](https://github.com/haskell/cabal/#conventions).
* [x] Any changes that could be relevant to users have been recorded in the changelog.
* [x] The documentation has been updated, if necessary.
* [x] If the change is docs-only, `[ci skip]` is used to avoid triggering the build bots.

Please also shortly describe how you tested your change. Bonus points for added tests!