@phadej 2017-12-13
I won't write amend user manual *yet*, as e.g. `cabal.project` parser doesn't benefit from the comma change yet (as it's still `ReadP`) based. And there are e.g. `constraints` where you'd like to use leading comma, IIRC.