@phadej 2017-12-13
ping @ezyang could you check *Implement availableSince* commit. I'll merge it in a week anyway, but if you want to comment on it before, there's time.