@23Skidoo 2017-12-19
> Surprisingly? this seems to be as performant as current master, parsing all Hackage packages starting with `a`: (including `acme-everything`:)

So `SPECIALIZED` makes it go a little bit slower, but use less memory? Interesting result.