@phadej 2017-12-19
The point of benchmark was to validate that the added abstraction doesn't add noticeable slowdown.