@23Skidoo 2017-12-19
Feel free to merge in the current state and tweak the design in a later separate PR (if you feel like it).