@phadej 2017-12-19
I'll add missing haddock and rebase when #4962 is merged.