@phadej 2017-12-25
Merged as #4971